Pick up filter docblock by copying it from the assignment node
If filters are documented like this:

  /**
   * A filter with some exciting documentation
   */
  $foo = apply_filters( 'filter', $foo );

then PHP-Parser associates the docblock with the assignment expression
instead of the function call. This means that the documentation isn't
actually retrieved by the hook reflector.

So, tweak the node visitor to check for assignment expressions that have
docblock and a filter as a subnode. The docblock can then be copied to
the subnode.