Add scribu/posts-to-posts to composer.json
minimum-stability set to "dev" because posts-to-posts requires scribu/scb-framework as dev-master. Dependency fails to load if parent project is not set to allow development branches.

Repository for posts-to-posts added manually because composer doesn't find it on its own.

Path for posts-to-posts manually set because type "wordpress-plugin" causes it to be installed in a wp-content directory created within this plugin directory.

Fixes #112