Decouple the importer class from WP-CLI
The `Importer` class was decoupled from WP-CLI except for its use as a
progress/error reporting function. I’ve fixed that by abstracting the
message reporting into two new methods, and implementing the use of
WP-CLI for reporting in a child class. We then use the child class for
importing.

See #82